replace custom errors with structured exceptions
